refactor: 💡 調整 buddhaSeven 相關變數名稱
