+++ src/models/buddhaSeven/buddhaSevenPeriods.model.ts
@@ -1,0 +1,12 @@
+/**
+ * @example {
+ *  "StartSevenDate": "2023-01-01",
+    "CompleteDate": "2023-01-07",
+    "Remarks":"前端新增測試"
+ * }
+ */
+export interface BuddhaSevenPeriod {
+  StartSevenDate: Date;
+  CompleteDate: Date;
+  Remarks?: string | null;
+}
