+++ src/models/index.ts
@@ -2,7 +2,7 @@ export * from "./lib/common.model";
 export * from "./lib/managers.model";
 export * from "./lib/rooms.model";
 export * from "./lib/users.model";
-export * from "./lib/buddhaSevenPeriods.model";
-export * from "./lib/buddhaSevenApply.model";
-export * from "./lib/buddhaSevenCheckIn.model";
+export * from "./buddhaSeven/buddhaSevenApplies.model";
+export * from "./buddhaSeven/buddhaSevenPeriods.model";
+export * from "./buddhaSeven/buddhaSevenCheckIns.model";
 export * from "./auth/jwt.model";
