--- src/models/lib/buddhaSevenApply.model.ts
@@ -1,10 +1,0 @@
-export interface BuddhaSevenApplyRequest {
-  UserId: number;
-  CheckInDate: Date;
-  CheckOutDate: Date;
-  CheckInDateBreakfast: boolean;
-  CheckInDateLunch: boolean;
-  CheckInDateDinner: boolean;
-  Remarks?: string;
-  UpdateUserId: number;
-}
